No OutletCollections in OS X, only iOS. So I read up on so-called advantageous, "bindings", which claim to be a lot less work and simpler than the usual glue code to tie view content to model data.
OMG. Having read the documentation on "bindings". it seems to me that they in fact require an order of magnitude MORE code (and complexity) than does simple, if not tedious glue code. I see no advantage to them what-so-ever. In fact, I think the entire "bindings" scheme is all but useless and frankly stupid. So I guess we'll stick with our simple, concise, so-called disadvantaged, "glue code". This glue code is simple and concise because it is much easier to see what is actually happening -- and there is in fact, at least in this case, a lot LESS "tedious" (and difficult to understand) code to write! *shrug*